[FVAC] First working implementation of natural policy gradient (NPG)
All 59 unit tests pass except the usual two FV tests in test_estimators_discretetime.py.

Goal: Estimate an optimal policy for ALL states, not only for the states in the optimal path found (e.g. think of labyrinth).
This would also avoid having an important subset of FV particles stuck at the same cell in a labyrinth, which happens because the policy of states *that are not visited very frequently* as an optimal policy is learned mimic the policy of adjacent states that are visited frequently. With a natural policy gradient approach, this problem should go away because the policy of each state is governed by the advantage function, which is correctly estimated by the Fleming-Viot estimator of the value functions.
All this was confirmed by the initial experiments run on FVAC and TDAC on a 2D gridworld with wind: the optimal policy now directs the agent in the right direction (with 100% probability in most cases)!

Note: The natural policy gradient approach uses a neural network simply as an implementation of action preferences using the softmax function. This means that the neural network is NOT used as a backpropagation network, i.e. its parameters are NOT updated by backpropagation, and therefore there is no requirement to compute the gradient of the parameters. This is why we use `requires_grad=False` when initializing the weights and biases of the network in PolNN.init_policy() so that parameters can be updated by setting DIRECTLY their values (as opposed to updating them using the parameters gradient, which is the common case in neural network learning). If `requires_grad=True` (the default), it is NOT possible to set the parameter values to a given value, as an error is issued ("RuntimeError: a leaf Variable that requires grad is being used in an in-place operation.").

As an aside: Added the requirement for torch in environment_rl.yml and fixed the name of the torch package in requirements_rl.txt.

DETAILS:
- learners/policies.py:
  - LeaActorCriticNN: new learn_natural() method that is the first implementation of the natural policy gradient (NPG) algorithm. So far, it does not seem to be producing satisfactory results.

- policies/parameterized.py:
  - PolNN:
    - init_policy(): Implemented the possibility of setting an initial policy value that *depends on the state* in the no-hidden-layer architecture (before this change, the initial policy value was the same for all states).
 This is useful when working with the natural policy gradient algorithm (NPG) to learn the policy, where the policy is proportional to the exponential of the advantage function.

    - new method set_policy_values() which calls init_policy() but it is a name that is more natural for the user when we just want to set the policy to a specific value during learning, which is the case when learning the policy using NPG.

- test/test_policies_parameterized.py: new test script that tests the PolNN class (neural network policy) when the neural network is defined with no hidden layer (NHL). See class Test_PolNN_NHL.

- test/test_estimators_models.py:
  - made assert_nn_model_parameter_values() a function that can be used by any other script, as it is useful to compare neural network model parameters used in assertions.
  - Test_Class_nn_backprop: added a test of the NNBackprop.set_params() method in the no-hidden-layer case.

- test/test_optimizers_discretetime.py:

- tests.py: (FVAC 2D)
  - Adapted the code for NPG, as follows:
	  - new variable `is_NPG` which is True when the neural network used to model the policy has no hidden layer.
	  - the learning rate for the optimizer is set to 10 for NPG (in which case it's not the optimizer learning rate but the learning rate of the NPG update formula).
	  - the LeaActorCriticNN.learn_natural() method is called when learning by NPG.
	  - added variables `is_NPG` and `nn_hidden_layer_sizes` to the list of saved objects into the pickle file.
	  - adapted the ALTOGETHER plots so that the long-run average reward is plotted instead of the episodic average reward when learning by NPG or when learning offline (as the episodic average reward is not computed in these cases). Now the plotted average reward is called `dict_R_toplot`.
  - I also modified the plotting process to refer to all `dict_*[learning_method]` objects instead of the `_all` objects so that we can plot different learning methods, not only the one last run.
  - Finally, in the FVAC 1D at the top of the file, I improved a few comments about the case when the policy is deterministic and improved the computation of the stationary distribution mu in that deterministic case so that all starting states are taken into account in its computation (see the calculation of `evector` which is now computed as a sum).