[Rebase] Rebase remote rl_labyrinth branch to local rl_labyrinth dated 04-Aug-2024
(I am not really sure what was done, as this is the first time I do a rebase... In principle doing a rebase means to reposition the changes done in the remote rl_labyrinth branch w.r.t. the code version at the *current* position of my local rl_labyrinth branch --which evolved since the remote rl_labyrinth branch had been updated.)

Goal: Be able to push my local rl_labyrinth branch to the remote repository, as in the remote repository the rl_labyrinth branch is the result of having merged the Alphonse branch into it, making push of my local rl_labyrinth branch to the remote repository impossible.

NOTE 1: Before starting this rebase I did a full copy of my local repository (i.e. the Python files, not the .git directory!) into ownCloud as directory `Python-v20240804-FVACWithNPG-BeforeMergeWithRemoteLabyrinthBranchResultingFromMergeWithAlphonse`.

NOTE 2: After I finished the conflict resolution and just before committing the changes I did a quick check of running the TDAC process with tests.py and it worked fine.
Note that I was not able to run the unit tests because the rebase process left a few temporary files among the test* files (e.g. test_optimizers_discretetime_BACKUP_1946.py) and this generated a test collection error

The rebase was started in SourceTree, using the option Merge, after having checked out my rl_labyrinth branch (positioned at the latest commit I had done locally) and then selecting the remote rl_labyrinth branch that I would like to merge. Finally I ticked the options appearing at the bottom of the Merge dialog box:
a) Create a new commit even if fast-forward is possible.
b) Rebase instead of MERGE.

I.e. the important option for performing a rebase is (b).

Before running the rebase process, I had to have my local repository clean, i.e. I had to commit all the changes (there were only two files to commit, the .gitignore file and the other one a PPT file that I stopped tracking).

During the rebase process, I got:
1) HEAD pointing at 145292b commit "[Bug] Fixed the Phi estimation..." from 21-Jul-2024, although the commit history line shown in SourceTree connected HEAD from the remote rl_labyrinth branch dated back to 21-Jun-2024 (yes, June not July)... So, I don't understand really what was going on.

2) A total of 8 files appeared as "changed" during the Rebase process, with two files having conflicts:
  - test_optimizers_discretetime.py
  - tests.py

3) I resolved the conflicts using the Merge Tool that appeared when clicking on "Resolve" (conflicts) in PyCharm (note that I switched to PyCharm because resolving the conflicts seemed much clearer there compared to SourceTree), where 3 columns appear: LEFT = my local version (which one I don't know precisely, but it said `rl_labyrinth` at the top of the column), CENTER = Result of the conflict resolution, RIGHT = the remote version of rl_labyrinth. I took the opportunity to improve the implementation that had been done by Alphonse in these conflicting files (e.g. defined new execution parameter `use_random_obstacles_set`, renamed seed_labyrinth --> seed_obstacles, improved logic regarding parameter `obstacles_set` in test_optimizers_discretetime.py, and so forth)
Note that while resolving conflicts I accepted a few lines from the left column (local copy) and a few lines from the right column (remote copy), depending on what was being merged).

4) I fixed an error that showed up in agents/learners/policies.py in the LeaActorCriticNN class where I had forgotten to add the `prob_states` parameter in the learn_offline_from_estimated_value_functions() function (so I added the parameter).

Note that the aforementioned 145292b commit had only one file changed (discrete.py), NOT 8 files as is the case now. But I think this has to do precisely with the fact that we are rebasing the remote rl_labyrinth branch, i.e. we are comparing the remote rl_labyrinth branch with a newer local commit... the question is which commit are we are comparing it with??