[MountainCar] Slightly improved discretization with capability of choosing different discretization scales for the position
Note: The originally named test_get_from_adjacent_states() (no renamed to no_test_get_from_adjacent_states()) in test_environments_mountaincars.py no longer works, as the discrete environment has been changed as described below. Since the test tests something that is actually not used in the simulation (theoretically computing the adjacent states to a state which was initially used to define the activation set for FV --but in the end the concept is incorrectly defined, so we should NOT use it for that purpose), it is no longer run (this is accomplished by renaming it to `no_test...`), but still kept (at least for now).

Goal: Be able to tune the number of discrete states used in the Mountain Car environment (both in velocity and position --so far only the velocity discretization could be tuned via constructor parameter `nv`) and thus be able to reduce the number of states to make the problem more manageable in the tabular scenario for value functions.

Note however that the position discretization tuning is not done directly, but via a factor that is used to adjust the force, gravity and max_speed parameters of the original mountaincar environment, which is continuous-state-valued. The new parameter is called `factor_for_force_and_gravity` which defaults to 20, giving 107 discrete positions. If we use 100 as factor, we get 21 discrete positions.
Essentially this factor is used to increase the value of the original force and gravity in order guarantee that there is (almost) always a transition to another discrete state when an action is applied. A reduced factor (say 0.25*factor_for_force_and_gravity) is used to adjust the max_speed parameter of the original mountaincar environment which controls how fast the car can go following the application of a force.

The slightly discretization improvement implies choosing the discrete state that is closer to the continuous-valued state generated by the mountain car dynamics after taking an action. The discretization is now equivalent to "rounding" the position and velocity to the grid, as opposed to truncating their values to the grid. See discretize() method.

This approach is NOT the most sensible one and will be soon change for an environment that performs the dynamics in the continuous-state space and THEN discretizes, if needed by the control problem to solve. This seems actually to be the way most people deal with the Mountain Car environment in the community!
Ex:
General: https://towardsdatascience.com/reinforcement-learning-applied-to-the-mountain-car-problem-1c4fb16729ba
Stanford: https://stanford-cs221.github.io/autumn2023/assignments/mountaincar/index.html
GitHub: https://github.com/noamsgl/Mountain-Car---Policy-Gradient/blob/main/main.py