[FVAC] NN models to approximate value functions, dropout option in policy model, execution on continuous- and discrete-state Mountain Car with 2 input neurons
This is almost the final commit among a series of partial commits so everything should work fine: out of the 63 unit tests in place, only 1 should fail:
test_estimators_discretetime.py::Test_EstDifferentialValueFunctions_EnvGridworld2DWithObstacles::test_Env_PolRandomWalk_MetFV

Note: The iterative plots generated by discrete.Simulator.run() when plot=True in tests.py were tested and work fine for both the Labyrinth environment and the discrete-state MountainCar environment. HOWEVER, they do NOT currently work for the continuous-state MountainCar environment because the plotting functions in the discrete.Simulator class are not prepared to handle function approximation yet. The fix is not so easy and it has to do with the TODO task added in ValueFunctionApproxNN.getValues() method in value_functions.py.

DETAILS:
- tests.py:
  - The setUpClass() call for the MountainCarDiscrete environment was completed so that it works fine for both the continuous-state and the discrete-state mountain car environments. Value function approximation is used for the continuous-state mountain car.

  - Added the option for dropout layers in the value functions (Mountain Car) and in the policy model (Labyrinth and Mountain Car).

  - Adapted the call to learner_ac.learn() for the case where value functions are approximated, for which we need to pass the learner (containing the NN models for V and Q) and the estimated expected reward, needed for the computation of the advantage function at each step used to updated the policy parameters when learning online.

  - Improved and fixed a few plots. Among other things, tried to make a sensible plot of the advantage function in the mountain car case which has many more states than the labyrinth environment.

  - Added the GIF plot showing the mountain car trajectory for the policy learned at the last policy learning replication.