[Plots] The ALTOGETHER plot in tests.py used to generate paper results conveys results better
I decided to show the paper results by plotting NOT the min/max curves but the +/- SE errors around the mean, AS WELL AS the median curve, in order to show the advantage of FVAC vs. TDAC in a cleaner and more convincing way.

These results are those presented in the final version of the EWRL-2024 paper.

Other improvements include:
- Showing the right axis in blue color, i.e. the same color as the curve whose values are plotted on the right axis, so that the reader more easily identifies where to read curve values.
- Unified the lower-end of the left axis to -1 percent of the maximum Y-axis value towards the negative values so that the scale is comparable between the case when the max attainable average reward is known or approximately known (e.g. 4x5 labyrinth with corridor) and when it is not known or difficult to compute (e.g. 6x8 labyrinth with random obstacles).
- The Y-axis label is now automatically generated depending on whether the normalized or the unnormalized average reward is plotted.