[FVAC] Fixed the unfair analysis of FVAC vs. TDAC where the absorption set A was the same for ALL replications!
Goal: Make a fairer FVAC vs. TDAC comparison by defining/estimating the absorption set A at the start of every replication even when estimate_absorption_set_at_every_step = False.
Note that when estimate_absorption_set_at_every_step = True, the estimation of the absorption set A was done at EVERY policy learning step so the described problem was not an issue in that case.

In addition (and this is related to the definition of the absorption set), I added a couple of commented out lines in discrete.py and in test_optimizers_discretetime.py which when uncommented can be used to estimate the absorption set A using the relative visit frequency as opposed to the *cumulative* relative visit frequency.
This is useful to reproduce results presented in the EWRL-2024 paper, which was based on the relative visit frequency, without having the hassle of going back to an earlier version just to perform this analysis change.

DETAILS:
- discrete.py:
  - Fixed a bug where an inexistent method get_index2d_from_index() was called.

- tests.py:
  - MAIN CHANGE: The absorption set is now estimated at the start of EVERY replication when estimate_absorption_set_at_every_step = False (i.e. when t_learn = 0), as opposed to using always the SAME absorption set estimated when creating `test_ac` which we already saw could be a very unlikely but favorable absorption set for FV, which is the responsible behind the EXCELLENT performance of FVAC vs. TDAC in the results presented in the camera-ready version of the EWRL-2024 paper! OMG, what did I do!! :(

  - Other minor changes, but possibly useful for increased functionality:
    - Option of selective absorption threshold (threshold_absorption_set) depending on the type of environment, either Gridworld or MountainCar.
    - Ad-hoc creation of an inverted L in the small 4x5 labyrinth in order to make reaching the exit more complicated, and thus perhaps observe more easily an advantage of FVAC. This is done between the "ENVIRONMENT" and "MODEL FOR POLICY" sections.
    - Added two commented out lines where we can set the default absorption set in case we want to compare a learning approach with another one, using specifically the same absorption set in both cases. This deemed useful when comparing learning with NPG vs. non-NPG because the random policy used at the start of the experiment varies a little bit but it suffices to perhaps obtain a significantly different absorption set (at least for the 6x8 labyrinth with random obstacles generated with seed 4217 presented in the EWRL-2024 paper.
    - Changed the M2 value (commonly called L in my notes and in the name of .pkl files containing results) to 30*N as opposed to 2*30*N (M2 = max_time_steps_fv_for_all_particles instead of M2 = 2*max_time_steps_fv_for_all_particles), just to make this parameter more transparent to me, the end user.
    - A few fixes and improvements about plotting results.