[Gridworlds] Improved EnvGridworld2D.plot() method so that the Start and Finish cells are plotted in pre-specified colors
Goal: Obtain the plot of the labyrinth as similar as possible to how I usually show it in presentations.

This implementation was not obvious but StackOverflow came to the rescue!

In test_optimizers_discretetime.py, the cells belonging to the absorption set are plotted using a Blues color palette with actual value range of proportion of visit between 0 and 1 (instead of between 0 and whatever the maximum proportion observed in the initial excursion was).
The goal of these two changes is that the cells appear as similar as possible to the cyan color used for the absorption set A in my presentations.
The first time I used this new plotting approach was in the preparation of my EWLR-2024 poster.