[Learning] Introduced the option of learning the policy using batches
IMPORTANT: Batch learning of the policy is logically NOT APPROPRIATE because the policy parameter changes after every batch (i.e. after every episode of the agent under the current policy) but the critic following those updates is NOT updated (i.e. the new policy is NOT re-evaluated), which is definitely wrong.
When tested, batch learning seems to work well, as learning VERY FAST from the onset, making the policy optimal in 4-5 steps (this corresponds to a basic test on a very small labyrinth of 3x4 with NO wind, both using TDAC and FVAC). However, once the optimal policy is attained and "learning" continues (because the policy learning process takes place for a fixed pre-specified number of steps), the policy is UNLEARNED... very quickly again... and what we observe is that the best action at one particular cell switches to "moving against an obstacle" with probability 1!

Note that batch learning mode defaults to False anyway.

In policies.py, the LeaActorCriticNN.learn() method has now a new parameter `batch_learning=False` which is used to handle both batch and non-batch learning.