[FV] Fixed the leakage that could occur from one experiment/replication to the next
Root cause of the leakage: Not all information related to the FV process was reset at the start of a new FV experiment run with discrete.Simulator, namely the information related to the absorption set and to the distribution used for the selection of start states for the E(T_A) simulation and for the FV simulation. This was currently done by hand by the user (e.g. see tests.py where, so far, I cumbersomely called _learner_value_functions_for_critic.setProbasStationaryStartStateET() and _learner_value_functions_for_critic.setProbasStationaryStartStateFV() to reset those distributions).
The situation was discovered by executing two FV tests in test_estimators_discretetime.py on the 2D gridworld with obstacles (one on the SOFT killing case, which at this point has not yet been committed --will come in the next commit), where the results of the second test (that uses SOFT killing) was different when it was run on its own and when it was run after the previous test on FV (which uses HARD killing).

This is a partial commit, so correct execution is not guaranteed.

DETAILS:
- discrete.py: Simulator class
  - _run_fv():
    - New reset of the learner done at the beginning of the method when `t_learn = 0`. This is needed to reset all information related to the FV learning which may be affected by previous executions of an FV learning, such as the absorption set or the distribution to use for the start state of the E(T_A) simulation and the FV simulation. This is crucial so that such information does NOT leak to a new experiment (replication) from a previously executed experiment.

- discrete/fv.py:
  - New attributes that allow resetting auxiliary information used by the FV simulation to their original values (at the time of construction), see private attributes such as `_absorption_set_at_construction` or `_probas_stationary_start_state_et_at_construction`.
  - New reset*() methods that reset attributes to their original value (at the time of object construction).

- tests.py: removed the "manual" and cumbersome reset of the distribution of start states for the estimation of E(T_A) and for the FV simulation, which is now done by discrete.Simulator._run_fv().