[FV] Simplified the particle reactivation process in _run_simulation_fv().reactivate_particle_internal()
In the HARD killing case, there is no way that the reactivated state is in the absorption set.
In the SOFT killing case, the assertion in discrete.Simulator._run_simulation_fv() that checks that the next state is NOT in the absorption set is disabled, because there is no real concept of absorption set in that case.

This is a partial commit, so correct execution is not guaranteed.