[FVAC] Improved estimation of E(T_A) by using estimates from previous policy learning steps; breakup of average reward into initial exploration, raw FV average, FV estimation
Summary:
a) For policy learning steps beyond the first one, the expected reabsorption time E(T_A) is estimated with the estimate of the previous policy learning step if no full reabsorption cycle was observed during the initial exploration.
b) The average reward observed during the initial exploration and the raw average reward observed during the FV simulation are now stored in the FV learner in the `average_reward_initial_exploration` and  the `average_reward_raw` attributes. These are reset by the LeaFV.reset() method whenever reset_average_reward=True.

Goals:
a) Use an estimate of E(T_A) that is expected to be a better estimate than the number T of steps run, which is probably a too extreme overestimation of E(T_A), when no full reabsorption cycles are observed. This strategy is based on the reasonable assumption that nor the policy nor the absorption set A change a lot between two consecutive policy learning steps. Of course, these assumptions may no longer hold if this estimate pooling strategy is pursued for several consecutive steps... but at the moment we don't check that.
b) Analyze how much the FV simulation contributes to the estimation of the long-run expected reward.

This is a partial commit, so correct execution is not guaranteed.

DETAILS:
- discrete.py: Simulator class
  - _estimate_value_functions_and_expected_reward_fv():
    - The estimated expected reabsorption time E(T_A) is now set to the value estimated at the previous policy learning step (for learning steps beyond the first one, i.e. t_learn > 0) when no reabsorption cycle has been observed during the initial exploration of the environment.

    - The average reward observed during the initial exploration (computed as the plain average of the observed rewards, and NOT by retrieving it from the average_reward attribute stored in the learner, which may contain a different value as it may be influenced by an initial estimate from the previous learning step) is now stored in the FV learner, so that we can analyze how much the FV simulation contributes to learning the long-run expected reward. This is done by a call to the LeaFV.setAverageRewardInitialExploration() method.

    - The raw average reward observed during the FV simulation is now stored in the FV learner by calling LeaFV.setAverageRewardRaw(), so that we get a feeling of how many rewards are observed by the FV particle system.

  - _run_simulation_fv():
    - Fixed the recording of rewards observed during the FV simulation, as the rewards observed after reactivation (e.g. when reactivation happens to a terminal state with reward) were currently not taken into account.

- tests.py:
  - New dictionaries to store the average reward breakdown:
    - dict_R_long_initial: average reward from initial exploration.
    - dict_R_long_fv_inflated: raw average reward by FV.
  - A new plot of the breakup of the average reward into its 3 newly recorded components for the last replication is now generated soon after the plot of the learning curve for the last replication.