[REFACTOR] Disentangled the trajectory update method from the average reward update method
The disentanglement affects only the GenericLearner class but NOT the Learner class that is specific to discrete-time discrete-state environments.
However, in the latter Learner class, defined in discrete/__init__.py, I have made clear that the _update_trajectory() method ALSO udpates the average reward by renaming the method to _update_trajectory_and_average_reward(). Note that I did NOT disentangle the two steps of updating the trajectory and the average reward in this case because: (i) the Learner._update_average_reward() doesn't receive any argument and it would not be clear how we update the average reward if we called this method from the outside world; (ii) the method is called from many other places (e.g. mc.py, td.py, etc.).

I also fixed one small bug, regarding printing messages in the console, that showed up when running the process (e.g. unit tests) with the option DEBUG_ESTIMATORS = True.

This is a partial commit, so correct execution is not guaranteed. However, all 67 unit tests should pass.

DETAILS:
- learners/__init__.py:
  - GenericLearner class: removed the call to update_average_reward() from the update_trajectory() method.

- learners/episodic/discrete.py:
  - Learner class: renamed method _update_trajectory() --> _update_trajectory_and_average_reward() in order to make clear that the method ALSO updates the average reward, not only the trajectory.

- simulators/__init__.py: renamed function update_trajectory() --> update_trajectory_and_average_reward(). The method now explicitly calls the GenericLearner.update_average_reward().

- simulators/queues.py: updated calls to the newly named update_trajectory_and_average_reward() function.

- learners/mc.py, td.py: updated calls to the newly named _update_trajectory_and_average_reward() method in the Learner class.