[FVAC] Fixed the CHEATY selection of FV start states as samples of the ACTIVATION set
Summary: The activation set should NOT be used as a set to select the start states of FV particles because we assume that the agent doesn't know anything about the environment dynamics, apart from what it learns from exploration. The current handling of the situation where no EXIT states from A were observed during the initial exploration of the environment was done in such a way that the activation set was used instead of the dictionary of states and their probabilities stored in the dist_proba_for_start_state passed to _run_simulation_fv() (which actually would include an alternative set for the start states when no EXIT states were observed; but this alternative set was not honoured).
Now, the only condition under which the activation set is used to select the FV start states is when such dictionary dist_proba_for_start_state is empty. But in that case, a WARNING message is now shown with label "CHEATING-WARNING:", generated by the Simulator._estimate_value_functions_and_expected_reward_fv() method in simulators/discrete.py.

This is a partial commit, so correct execution is not guaranteed.