[FVAC] Fixed the estimation of the absorption set at the first learning step
Goal: Make implementation more clear and avoid the error that the absorption set is tagged as having shrinked at the very first learning step of a new replication, which does not make sense because a new replication must compute the absorption set from scratch (even in the case when the user requests NOT to the update the absorption set during learning, as in that case we need to start with an absorption set identified at the very beginning, and this absorption set may be allowed to change at every replication --recall the very fast and INCORRECT learning by FVAC that was reported in the EWRL-2024 paper just because the absorption set identified during the test_ac class construction in tests.py --used subsequently in all replications (incorrectly)-- was very favorable for FV).

This change implied removing a section in tests.py (for now disabled inside an `if False` block so that I remember in the future why that step disappeared from the testing process and I don't spend time re-establishing it again) which decied that the absorption set A should be estimated at the start of every new replication. This is now responsibility of the implementation in discrete.Simulator, as it should be (i.e. it cannot be the responsibility of the user to remember this!).

I've also added an informative message to the user about the logic behind the estimated expected reabsorption time E(T_A), due to different situations during the initial exploration that is run to estimate it. This change was triggered by the fact that in the log files I've seen E(T_A) taking a real value when NO cycles were observed during the initial exploration, which doesn't make sense because in that case the value of E(T_A) is set to the total simulation time, which is discrete. The reason for those real values is that, when no cycles are observed, the E(T_A) is set to the estimate computed at the PREVIOUS learning step (where probably more than one cycle has been observed).

This is a partial commit, so correct execution is not guaranteed.