[FVAC] Fixed one detail about the selection of FV start states and better inform the analysis of states visited outside A
The fixed small detail has to do with the condition on `less_frequently_visited_states_case[:1] != "3"` in _estimate_value_functions_and_expected_reward_fv() in simulators/discrete.py, which was now completely deactivated (i.e. it is no longer possible to enter that block because it doesn't make sense to use backup states, as they are more likely NOT a subset of the boundary of A. Hence, the FV particles could start at ANY state outside A and this generates problems in the policy learning process.

This is a partial commit, so correct execution is not guaranteed.