[FVAC] Implemented two new FV stopping criteria with parametric estimation of P(T>t) and fixed t_learn handling
Goal: Leverage information obtained after the last absorption time collected for the N particles about the conditional occupation probability Phi(x,t) which would be able to contribute perhaps significantly to the FV integral.

Summary: A new enum called StoppingCriterion was defined for the FV simulation. The enum can take 4 possible values as described in the documentation: two of them use a parametric estimation of the survival probability distribution P(T>t), and the other two the usual moments estimator.
The two new strategies that use a parametric estimation of the survival probability are encoded as enum values MAX_TIME_STEPS and MAX_TIME_STEPS_OR_MIN_PROP_ABSORBED_PARTICLES_WHATEVER_COMES_LATER.
The latter was created to leverage as much information as possible from the FV particle system, as the simulation stops when the latest out of the two conditions that could sensibly stop the simulation (either achieving a maximum budget of steps or the absorption of the N particles at least once) occurs.

The parametric estimation of the survival probability implies: (i) P(T>t) is estimated as exp(-alpha*t) where alpha = 1/E(T), the exponential killing rate estimated from the observed first-time survival times, (ii) the FV integral and average reward is computed at the END of the FV simulation, as opposed to iteratively.

In addition, a few fixes were applied following the changes committed on 30-May-2025 where the absorption set A is estimated at the first policy learning step, when t_learn = 0, in order to discharge the user from the responsibility of requiring to re-estimate the absorption set at the start of a new replication.
The solution was to change the default value of t_learn in methods to t_learn=-1, to indicate that by default FV simulations are NOT run as part of a policy learning process, as is the case when running unit tests. Otherwise, all FV unit tests in test_estimators_discretetime.py would fail.
This change is also linked with a different handling of the t_learn value to decide the reset of the FV learner: from `t_learn == 0` --> `t_learn <= 0` so that reset also happens at the default value of t_learn=-1, as this should be the case when running unit tests in order to avoid that unit tests run consecutively give different results compared to those given when run individually, just because a proper reset of the learner was not carried out. This happened precisely with the Soft Killing test in Test_EstDifferentialValueFunctions_EnvGridworld2DWithObstacles, `test_Env_PolRandomWalk_MetFV_SoftKilling()` because the test shares the same learner as the test coming before it (which is executed first), `test_Env_PolRandomWalk_MetFV()`.

This is a partial commit, so correct execution is not guaranteed.
However all 68 unit tests should pass.

DETAILS:
- estimators/fv.py:
  - New function estimate_proba_stationary_parametric() used to estimate the stationary probability using a parametric estimation of the survival probability P(T>t).
  This requires a separate function from the case when P(T>t) is estimated non-parametrically because no merge between the step function P(T>t) and the step function Phi(t,x) is required in this case.

- simulators/discrete.py: Simulator class
  - Changed the default proportion of absorbed particles once from 0.9 to 1.0 because the user should not expect to have a default value that is "not standard" or not expected.

  - [IMPORTANT] Two changes performed on the `t_learn` parameter, in terms of handling and its default value:
    - Default value for t_learn: was changed from t_learn=0 --> t_learn=-1 in order to indicate that by default we are NOT running the simulation as part of a policy learning process --except for one t_learn parameter in _setup_plots() because this function has been designed specifically for the policy learning context.
    This is needed because the new implementation of the policy learning process is such that the absorption set A is ALWAYS estimated at the first policy learning step, in order to guarantee that the absorption set is estimated from scratch at the start of a new replication (without making the user worry about this --see commit from 30-May-2025).
    However, we do NOT want this to happen when we run unit tests like those written in test_estimators_discretetime.py because, in those tests, the absorption set is defined and fixed at the start and should therefore NOT be estimated at the beginning of the simulation.

    - Handling of t_learn: the condition for resetting the learner was changed from `t_learn == 0` --> `t_learn <= -1` so that resetting also happens for the newly default value used for t_learn as t_learn=-1. Otherwise, unit tests could fail when run one after the other because a proper reset of the learner was not done before starting the second test (when they share the same learner).

  - _run_simulation_fv():
    - Implemented the handling all 4 stopping criteria that are now available by the new StoppingCriterion enum defined in simulators/fv.py.

    - When DEBUG_ESTIMATORS = True, added a plot of the parametric estimation of P(T>t) on top of the non-parametric estimation, for comparison purposes, only when the stopping criterion is NOT one of those requesting the parametric estimation (as o.w. these plots are generated by the estimate_stationary_probabilities() function defined in estimators/fv.py.

  - Deprecated other _run_simulation_fv*() functions (e.g. _run_simulation_fv_fraiman()) because they do not longer follow the same signature as _run_simulation_fv() after the stopping criterion for FV added flexibility.

- simulators/fv.py: New StoppingCriterion enum to define the stopping criterion for the FV simulation. See Summary above and the documentation of the class for more details.

- test_estimators_discretetime.py: Changes related to the new StoppingCriterion enum. Further checks on the consistency of the state counts and the number of simulation steps run were added.

- test_FV.py: new test_parametric_estimation_of_FV_integral() method to test the parametric estimation of the survival probability distribution.

- tests.py:
  - Adapted the script to the new StoppingCriterion enum.
  - Due to the new stopping criteria available for Fleming-Viot, the value of M2 was increased from being equal to M1 to 2*M1 so that there is a time during which the FV simulation can be stopped by a less than 100% absorption of the FV particles, if that should be requested (I recall that before M1, even if that condition happens, the FV simulation is not stopped because M1 is considered to be the minimum time during which the FV simulation should run.